Time Spent variables should be of type float, not int

+++ TIMEDSORTS/HeapSortdef.h
@@ -22,7 +22,7 @@ private:
 	int comparisons;
 	clock_t h_timeS;			//start clock
 	clock_t h_timeE;			//end clock
-	double h_time_Spent;
+	float h_time_Spent;
 };
 
 HeapSort::HeapSort()
